refactor(app-tools): reuse copy upload config when using Rspack
