+++ .changeset/spotty-chairs-design.md
@@ -1,0 +1,7 @@
+---
+'@modern-js/app-tools': patch
+---
+
+refactor(app-tools): reuse copy upload config when using Rspack
+
+refactor(app-tools): 使用 Rspack 时复用 copy upload 配置
