+++ packages/solutions/app-tools/src/builder/builder-rspack/index.ts
@@ -6,7 +6,6 @@ import { builderPluginAdpaterCopy } from './adapterCopy';
 export function createRspackBuilderForModern(
   options: BuilderOptions<'rspack'>,
 ) {
-  // Todo Copy config/upload
   return generateBuilder(options, builderRspackProvider, {
     modifyBuilderInstance(builder) {
       builder.addPlugins([builderPluginAdpaterCopy(options)]);
