+++ packages/solutions/app-tools/src/builder/generator/index.ts
@@ -3,11 +3,7 @@ import type {
   BuilderInstance,
 } from '@modern-js/builder-shared';
 import { createBuilder } from '@modern-js/builder';
-import {
-  BuilderOptions,
-  ModifyBuilderConfig,
-  ModifyBuilderInstance,
-} from '../shared';
+import { BuilderOptions, ModifyBuilderInstance } from '../shared';
 import { Bundler } from '../../types';
 import { createBuilderProviderConfig } from './createBuilderProviderConfig';
 import { getBuilderTargets } from './getBuilderTargets';
@@ -18,25 +14,23 @@ export type GenerateProvider = (c: { builderConfig: any }) => BuilderProvider;
 /**
  * @param options BuilderOptions
  * @param generateProvider GenerateProvider
- * @param utils - ModifyBuilderConfig, ModifyBuilderInstance
+ * @param utils - ModifyBuilderInstance
  * @returns BuilderInstance
  */
 export async function generateBuilder<B extends Bundler>(
   options: BuilderOptions<B>,
   generateProvider: GenerateProvider,
   utils?: {
-    modifyBuilderConfig?: ModifyBuilderConfig<B>;
     modifyBuilderInstance?: ModifyBuilderInstance;
   },
 ) {
   const { normalizedConfig, appContext } = options;
-  const { modifyBuilderConfig, modifyBuilderInstance } = utils || {};
+  const { modifyBuilderInstance } = utils || {};
 
   // create provider
   const builderConfig = createBuilderProviderConfig<B>(
     normalizedConfig,
     appContext,
-    modifyBuilderConfig,
   );
 
   const provider = generateProvider({
