+++ packages/solutions/app-tools/src/types/config/html.ts
@@ -1,7 +1,9 @@
-import type { BuilderConfig } from '@modern-js/builder-webpack-provider';
-import type { BuilderConfig as RsBuilderConfig } from '@modern-js/builder-rspack-provider';
 import { UnwrapBuilderConfig } from '../utils';
+import type {
+  WebpackBuilderConfig,
+  RspackBuilderConfig,
+} from '../../builder/shared';
 
 export type { SharedHtmlConfig } from '@modern-js/builder-shared';
-export type HtmlUserConfig = UnwrapBuilderConfig<BuilderConfig, 'html'>;
-export type RsHtmlUserConfig = UnwrapBuilderConfig<RsBuilderConfig, 'html'>;
+export type HtmlUserConfig = UnwrapBuilderConfig<WebpackBuilderConfig, 'html'>;
+export type RsHtmlUserConfig = UnwrapBuilderConfig<RspackBuilderConfig, 'html'>;
