+++ packages/solutions/app-tools/src/types/config/output.ts
@@ -1,12 +1,17 @@
-import type { BuilderConfig } from '@modern-js/builder-webpack-provider';
-import type { BuilderConfig as RsBuilderConfig } from '@modern-js/builder-rspack-provider';
 import type { SharedOutputConfig as BuilderSharedOutputConfig } from '@modern-js/builder-shared';
 import type { SSGConfig } from '@modern-js/types';
+import {
+  WebpackBuilderConfig,
+  RspackBuilderConfig,
+} from '../../builder/shared';
 import { UnwrapBuilderConfig } from '../utils';
 
-export type BuilderOutputConfig = UnwrapBuilderConfig<BuilderConfig, 'output'>;
+export type BuilderOutputConfig = UnwrapBuilderConfig<
+  WebpackBuilderConfig,
+  'output'
+>;
 export type RsBuilderOutputConfig = UnwrapBuilderConfig<
-  RsBuilderConfig,
+  RspackBuilderConfig,
   'output'
 >;
