+++ packages/solutions/app-tools/src/types/config/source.ts
@@ -1,9 +1,11 @@
-import type { BuilderConfig } from '@modern-js/builder-webpack-provider';
-import type { BuilderConfig as RsBuilderConfig } from '@modern-js/builder-rspack-provider';
 import type { SharedSourceConfig as BuilderSharedSourceConfig } from '@modern-js/builder-shared';
+import type {
+  WebpackBuilderConfig,
+  RspackBuilderConfig,
+} from '../../builder/shared';
 
-export type BuilderSourceConfig = NonNullable<BuilderConfig['source']>;
-export type RsBuilderSourceConfig = NonNullable<RsBuilderConfig['source']>;
+export type BuilderSourceConfig = NonNullable<WebpackBuilderConfig['source']>;
+export type RsBuilderSourceConfig = NonNullable<RspackBuilderConfig['source']>;
 
 export type Entry =
   | string
